xml rpc methods
git-svn-id: https://svn.code.sf.net/p/osirix/code/osirix@2903 4e9e6e3d-4551-47eb-9dbe-a6127f9400a1

+++ PluginManager.m
@@ -644,7 +644,7 @@ PluginManager			*pluginManager = 0L;
 
 #pragma mark plugins
 
-int sortPluginArray(id plugin1, id plugin2, void *context)
+NSInteger sortPluginArray(id plugin1, id plugin2, void *context)
 {
     NSString *name1 = [plugin1 objectForKey:@"name"];
     NSString *name2 = [plugin2 objectForKey:@"name"];
